Revert manual merging to figure out, what's the problem
