+++ datalad/tests/test_utils.py
@@ -1,0 +1,31 @@
+# emacs: -*- mode: python; py-indent-offset: 4; tab-width: 4; indent-tabs-mode: nil -*-
+# ex: set sts=4 ts=4 sw=4 noet:
+# ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
+#
+#   See COPYING file distributed along with the datalad package for the
+#   copyright and license terms.
+#
+# ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
+"""Test testing utilities
+
+"""
+
+import os
+
+from nose.tools import ok_, assert_false
+from mock import patch
+
+from datalad.tests.utils import with_tempfile
+
+
+@with_tempfile
+def _with_tempfile_decorated_dummy(path):
+    return path
+
+
+def test_with_tempfile_dir_via_env_variable():
+    target = os.path.join(os.path.expanduser("~"), "dataladtesttmpdir")
+    assert_false(os.path.exists(target), "directory %s already exists." % target)
+    with patch.dict('os.environ', {'DATALAD_TESTS_TEMPDIR': target}):
+        filename = _with_tempfile_decorated_dummy()
+        ok_(filename.startswith(target))
